7일차

+++ src/main/webapp/jsp/article/detail.jsp
@@ -10,7 +10,7 @@
 <html>
 <head>
 <meta charset="UTF-8">
-<title>Article Detail</title>
+<title>게시글 상세보기</title>
 <style type = "text/css">
 	a {
 		color : inherit;
@@ -32,8 +32,12 @@
 </head>
 <body>
 
-	<h1>Article Detail</h1>
+	<h2>게시글 상세보기</h2>
 
+	<%@ include file="/jsp/common/topBar.jsp" %>
+	
+	<br/>
+	
 	<nav class="menu-bar">
 		<ul>
 			<li><h2><%= article.get("id") %>번 게시글</h2></li>
@@ -42,19 +46,17 @@
 			<li>수정일시 : <%= article.get("updateDate") %></li>
 			<li>글 제목 : <%= article.get("title") %></li>
 			<li>글 내용 : <%= article.get("body") %></li>
-			<li>작성자 : <%= article.get("writerId") %></li>
+			<li>작성자 : <%= article.get("memberId") %></li>
 		</ul>
 	</nav>
 	
 	<br>
-	
-	
-	<button type="button" onclick="location.href='modify?id=<%= article.get("id") %>'">글 수정하기</button>
-	<button type="button" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false; location.href='delete?id=<%= article.get("id") %>'" >글 삭제하기</button>
+		
+	<% if (loginMemberNumber == (int) article.get("memberNumber")) { %>
+		<button type="button" onclick="location.href='modify?id=<%= article.get("id") %>'">글 수정하기</button>
+		<button type="button" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false; location.href='delete?id=<%= article.get("id") %>'" >글 삭제하기</button>
+	<% } %>
 
-<!-- 	
-<button type="button" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false; location.href='delete?id=<%= article.get("id") %>'" >글 삭제하기</button>
--->
 	<button type="button" onclick="location.href='list'">게시글 목록</button>
 	
 </body>
